Fix Groovy/Java joint compilation by stashing files from Java compilation called by Groovy compiler
