+++ subprojects/language-java/src/main/java/org/gradle/api/internal/tasks/compile/incremental/recomp/JavaRecompilationSpecProvider.java
@@ -18,10 +18,16 @@ package org.gradle.api.internal.tasks.compile.incremental.recomp;
 
 import org.gradle.api.file.FileTree;
 import org.gradle.api.internal.file.FileOperations;
+import org.gradle.api.internal.tasks.compile.JavaCompileSpec;
+import org.gradle.api.internal.tasks.compile.incremental.compilerapi.deps.GeneratedResource;
+import org.gradle.api.internal.tasks.compile.incremental.transaction.CompileTransaction;
+import org.gradle.api.internal.tasks.compile.incremental.transaction.JavaCompileTransaction;
+import org.gradle.api.tasks.util.PatternSet;
 import org.gradle.internal.file.Deleter;
 import org.gradle.work.FileChange;
 
 import java.util.Collections;
+import java.util.Map;
 import java.util.Set;
 
 public class JavaRecompilationSpecProvider extends AbstractRecompilationSpecProvider {
@@ -44,5 +50,10 @@ public class JavaRecompilationSpecProvider extends AbstractRecompilationSpecProv
     @Override
     protected boolean isIncrementalOnResourceChanges(CurrentCompilation currentCompilation) {
         return currentCompilation.getAnnotationProcessorPath().isEmpty();
+    }
+
+    @Override
+    protected CompileTransaction newCompileTransaction(JavaCompileSpec spec, RecompilationSpec recompilationSpec, PatternSet classesToDelete, Map<GeneratedResource.Location, PatternSet> resourcesToDelete, FileOperations fileOperations, Deleter deleter) {
+        return new JavaCompileTransaction(spec, classesToDelete, resourcesToDelete, fileOperations, deleter);
     }
 }
